+++ sdk/storage/azure-storage-blobs/test/ut/page_blob_client_test.cpp
@@ -99,21 +99,32 @@ namespace Azure { namespace Storage { namespace Test {
     auto downloadContent = pageBlobClient.Download();
     EXPECT_EQ(ReadBodyStream(downloadContent.Value.BodyStream), blobContent);
 
-    auto pageRanges = pageBlobClient.GetPageRanges().Value;
-    EXPECT_TRUE(pageRanges.ClearRanges.empty());
-    ASSERT_FALSE(pageRanges.PageRanges.empty());
-    EXPECT_EQ(static_cast<uint64_t>(pageRanges.PageRanges[0].Offset), 3_KB);
-    EXPECT_EQ(static_cast<uint64_t>(pageRanges.PageRanges[0].Length.Value()), 3_KB);
+    std::vector<Core::Http::HttpRange> pageRanges;
+    for (auto pageResult = pageBlobClient.GetPageRanges(); pageResult.HasMore();
+         pageResult.NextPage(Azure::Core::Context()))
+    {
+      pageRanges.insert(
+          pageRanges.end(), pageResult.PageRanges.begin(), pageResult.PageRanges.end());
+    }
+    ASSERT_FALSE(pageRanges.empty());
+    EXPECT_EQ(static_cast<uint64_t>(pageRanges[0].Offset), 3_KB);
+    EXPECT_EQ(static_cast<uint64_t>(pageRanges[0].Length.Value()), 3_KB);
 
     Azure::Storage::Blobs::GetPageRangesOptions options;
     options.Range = Core::Http::HttpRange();
     options.Range.Value().Offset = 4_KB;
     options.Range.Value().Length = 1_KB;
-    pageRanges = pageBlobClient.GetPageRanges(options).Value;
-    EXPECT_TRUE(pageRanges.ClearRanges.empty());
-    ASSERT_FALSE(pageRanges.PageRanges.empty());
-    EXPECT_EQ(static_cast<uint64_t>(pageRanges.PageRanges[0].Offset), 4_KB);
-    EXPECT_EQ(static_cast<uint64_t>(pageRanges.PageRanges[0].Length.Value()), 1_KB);
+    pageRanges.clear();
+    for (auto pageResult = pageBlobClient.GetPageRanges(options); pageResult.HasMore();
+         pageResult.NextPage(Azure::Core::Context()))
+    {
+      pageRanges.insert(
+          pageRanges.end(), pageResult.PageRanges.begin(), pageResult.PageRanges.end());
+    }
+    EXPECT_TRUE(pageRanges.empty());
+    ASSERT_FALSE(pageRanges.empty());
+    EXPECT_EQ(static_cast<uint64_t>(pageRanges[0].Offset), 4_KB);
+    EXPECT_EQ(static_cast<uint64_t>(pageRanges[0].Length.Value()), 1_KB);
 
     auto snapshot = pageBlobClient.CreateSnapshot().Value.Snapshot;
     // |_|_|_|x|  |x|x|_|_| This is what's in snapshot
@@ -123,13 +134,22 @@ namespace Azure { namespace Storage { namespace Test {
     pageBlobClient.ClearPages({3_KB, 1_KB});
     // |x|_|_|_|  |x|x|_|_|
 
-    pageRanges = pageBlobClient.GetPageRangesDiff(snapshot).Value;
-    ASSERT_FALSE(pageRanges.ClearRanges.empty());
-    ASSERT_FALSE(pageRanges.PageRanges.empty());
-    EXPECT_EQ(pageRanges.PageRanges[0].Offset, 0);
-    EXPECT_EQ(static_cast<uint64_t>(pageRanges.PageRanges[0].Length.Value()), 1_KB);
-    EXPECT_EQ(static_cast<uint64_t>(pageRanges.ClearRanges[0].Offset), 3_KB);
-    EXPECT_EQ(static_cast<uint64_t>(pageRanges.ClearRanges[0].Length.Value()), 1_KB);
+    pageRanges.clear();
+    std::vector<Core::Http::HttpRange> clearRanges;
+    for (auto pageResult = pageBlobClient.GetPageRangesDiff(snapshot); pageResult.HasMore();
+         pageResult.NextPage(Azure::Core::Context()))
+    {
+      pageRanges.insert(
+          pageRanges.end(), pageResult.PageRanges.begin(), pageResult.PageRanges.end());
+      clearRanges.insert(
+          clearRanges.end(), pageResult.ClearRanges.begin(), pageResult.ClearRanges.end());
+    }
+    ASSERT_FALSE(pageRanges.empty());
+    ASSERT_FALSE(pageRanges.empty());
+    EXPECT_EQ(pageRanges[0].Offset, 0);
+    EXPECT_EQ(static_cast<uint64_t>(pageRanges[0].Length.Value()), 1_KB);
+    EXPECT_EQ(static_cast<uint64_t>(clearRanges[0].Offset), 3_KB);
+    EXPECT_EQ(static_cast<uint64_t>(clearRanges[0].Length.Value()), 1_KB);
   }
 
   TEST_F(PageBlobClientTest, UploadFromUri)